+++ android/app/src/main/java/com/stocknotifier/MainApplication.java
@@ -3,6 +3,10 @@ package com.stocknotifier;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
+import com.horcrux.svg.SvgPackage;
+import io.invertase.firebase.RNFirebasePackage;
+import io.invertase.firebase.messaging.RNFirebaseMessagingPackage;
+import io.invertase.firebase.notifications.RNFirebaseNotificationsPackage;
 import com.reactnativecommunity.asyncstorage.AsyncStoragePackage;
 import com.oblador.vectoricons.VectorIconsPackage;
 import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
@@ -26,10 +30,15 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-	  new AsyncStoragePackage(),
-	  new VectorIconsPackage(),
-	  new RNGestureHandlerPackage(),
-	  new ClientLoginPackage());
+	        new ClientLoginPackage());
+          new SvgPackage(),
+          new RNFirebasePackage(),
+          new RNFirebaseMessagingPackage(),
+          new RNFirebaseNotificationsPackage(),
+          new AsyncStoragePackage(),
+          new VectorIconsPackage(),
+          new RNGestureHandlerPackage()
+      );
     }
 
     @Override