+++ app/Components/TickerCard.js
@@ -1,17 +1,41 @@
 import React, { Component } from 'react';
 import {
-  Text, View, TouchableOpacity, Alert, StyleSheet, Dimensions, 
+  Text, View, TouchableOpacity, Alert, StyleSheet, Dimensions,
 } from 'react-native';
 import Ionicon from 'react-native-vector-icons/Ionicons';
 
+const API_URL = 'http://cs130-stock-notifier-http-server.us-west-1.elasticbeanstalk.com/add_ticker';
+
 export default class TickerCard extends Component {
 
-  onAddPress(name) {
-    Alert.alert(name + " was added!")
+  async onAddPress(name) {
+    // TODO: load this stuff in from async storage
+    const body = {
+      username: 'brian',
+      session_id: '0747f11d56c721d49fae05e96abf261c192503b6bf59a2897f069738c3cbe8ff',
+      tickers: [name],
+    };
+
+    // Fetch the data from the API 
+    await fetch(API_URL, {
+      method: 'POST',
+      body: JSON.stringify(body),
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    })
+      .then((response) => { return response.json(); })
+      .then((data) => {
+        Alert.alert(name + ' was added!');
+      })
+      .catch((error) => {
+        console.log(error);
+        Alert.alert(name + ' failed to be added at this time');
+      });
   }
 
   render() {
-    let { id, name, price } = this.props;
+    const { id, name, price } = this.props;
 
     // No price means it's just the list for Add Tickers
     if (price === 'ADD') {
@@ -30,11 +54,11 @@ export default class TickerCard extends Component {
     else {
       return (
         <View style={styles.card}>
-          <View style={styles.left}>
+          <View>
             <Text style={styles.idText}>{id}</Text>
             <Text style={styles.nameText}>{name}</Text>
           </View>
-          <View style={styles.right}>
+          <View>
             <Text style={styles.priceText}>${price}</Text>
           </View>
         </View>
@@ -75,5 +99,6 @@ const styles = StyleSheet.create({
   },
   priceText: {
     fontSize: 20,
+    color: 'black',
   },
 });